do not allow old data replacing newer routing cache entries
An the follower of the noble principle of flooding was brought before the
king of the Node as it was a custom in these lands that all travelers passing
through answered before the king telling the tales of lands far away. He told
the king of his journeys and what evil had he encountered. Then the king spoke:
"Alas! If only thou hast come here swifter I would not have my men all over the
graph looking for what thou already know!". The guest wanted to describe the
perils that awaited him on his way to the Node, but before he could do that the
king continued: "Strange are your words though for different news I had heard
from my men who traveled through the graph. Nay! I am not going to take counsel
from you." Grat sorrow these words brought to the heart of the messanger,
sorrow that remained till the day his dead body was collected from the streets
of the Node.